@@ -5,7 +5,7 @@ homepage: https://www.yeradis.com
 author: Yeradis P. Barbosa Marrero <yeradis@gmail.com>
 
 environment:
-  sdk: '>=2.5.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 #dependencies:
 #  path: ^1.4.1
